refactor(detail): extract FadingTitleScaffold shared component
Factor the duplicated scroll-driven title fade-in logic (state, Scaffold,
SimpleTopBar and the LayoutCoordinates plumbing) into a reusable
FadingTitleScaffold in core. Magical item and spell detail screens now
delegate to it, and future detail screens can reuse the same behaviour.
